+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - JDK17 is the minimum requirement
+- replaced references to "Alphanum" with "Natural Sort Order" as the original website is gone
 
 ### Fixed